Move MouseWheelAnimation to Mapsui.Animations
