+++ Mapsui/INavigator.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using Mapsui.Animations;
-using Mapsui.UI;
 using Mapsui.Utilities;
 
 namespace Mapsui;
